Update "conform" documentation for lat/lon.
For every type we understand but CSV, the lat/lon is implicitly named
by being in the geometry field, so the contributer doesn't need to
specify it. The Python code will ignore these settings. It's possible
the Node code relies on them being set; if so, the Node code should
assume the names "X" and "Y" to match the rest of how it works.

For details on this change, see https://github.com/openaddresses/machine/issues/73