Update Kent County, MI
As shown in the [build output](http://data.openaddresses.io.s3.amazonaws.com/runs/33309/output.txt),
it doesn't look like the current URL works: it was run through some
proxy which hopefully isn't needed.

This also changes the URL from an ESRI layer that has only address info,
back to one that has full parcel data, as well as super valuable
zipcodes.

This URL was added in PR #359 but then changed in #1244, so if there was
a reason we can change it back. But it looks like this will give full
polygon and address data, where the other won't.