Likely better source for Pope
I'm thinking this is a better long-term source for Pope because a) it's found via their open data site and b) it doesn't have "TEST" in the URL. However, we lose the unit information. I queried the current source and all the records had UNITID as null. That's why I think that the unit info was withheld from the open data site.