Better source for Becker (#4495)
* Better source for Becker

This driveway point source appears to be better. The BASENAME field only has 11 records with nulls, and ADDRESS has 15, some of which otherwise appear to have real data. So this field mapping looks like an upgrade over using ADDRESS with a postfixed_street function.

* Even better Becker source

This shapefile download is from their open data delivery site, which makes more sense to use than scraping their REST endpoint. Unfortunately there's still a record in here with a null geometry, but I'll be asking them to fix that.